% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigFeatures.R
\name{bigFeatures}
\alias{bigFeatures}
\title{Create bigFeatures object}
\usage{
bigFeatures(matrix, rowData = NULL, colData = NULL)
}
\arguments{
\item{matrix}{Numeric matrix with samples on rows and features on columns}

\item{rowData}{data.frame with sample metadata. If NULL, uses rownames to create sample_name column}

\item{colData}{data.frame with feature metadata. If NULL, uses colnames to create feature_name column}
}
\value{
A bigFeatures object (list with elements: features, rowData, colData)
}
\description{
A list-based S3 class for storing feature data (genes/transcripts) in FBM format
with samples on rows and features on columns.
}
