% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigQTL-package.R
\docType{package}
\name{bigQTL-package}
\alias{bigQTL-package}
\title{bigQTL: Scalable Conditional QTL Analysis with File-Backed Matrices}
\description{
Performs conditional cis-QTL analysis (stepwise conditioning and all-but-one) using file-backed genotype and phenotype matrices from the bigstatsr/bigsnpr ecosystem. Includes eigenMT multiple testing correction and data preparation utilities (genotype PCA, phenotype PCA, RINT transformation).
}
\author{
\strong{Maintainer}: Pierre-Paul Axisa \email{pp.axisa@gmail.com}

Authors:
\itemize{
  \item Pierre-Paul Axisa \email{pp.axisa@gmail.com}
}

}
\keyword{internal}
