% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_feature_indices}
\alias{get_feature_indices}
\title{Match feature names to indices in a bigFeatures object}
\usage{
get_feature_indices(bigfeatures, feature_names)
}
\arguments{
\item{bigfeatures}{bigFeatures object with \code{$colData} containing
a \code{feature_name} column}

\item{feature_names}{Character vector of feature names to look up}
}
\value{
Integer vector of column indices into the feature FBM
}
\description{
Match feature names to indices in a bigFeatures object
}
