% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_pheno_indices}
\alias{get_pheno_indices}
\title{Match phenotype names to indices in a bigPheno object}
\usage{
get_pheno_indices(bigpheno, pheno_names)
}
\arguments{
\item{bigpheno}{bigPheno object with \code{$colData} containing
a \code{pheno_name} column}

\item{pheno_names}{Character vector of phenotype names to look up}
}
\value{
Integer vector of column indices into the phenotype FBM
}
\description{
Match phenotype names to indices in a bigPheno object
}
