% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_snp_indices}
\alias{get_snp_indices}
\title{Match SNP names to indices in a bigSNP object}
\usage{
get_snp_indices(bigsnp, snp_names)
}
\arguments{
\item{bigsnp}{bigSNP object with \code{$map} containing
a \code{marker.ID} column}

\item{snp_names}{Character vector of SNP names to look up}
}
\value{
Integer vector of column indices into the genotype FBM
}
\description{
Match SNP names to indices in a bigSNP object
}
