% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenMT.R
\name{lw_shrink_cor}
\alias{lw_shrink_cor}
\title{Apply Ledoit-Wolf shrinkage to a correlation matrix}
\usage{
lw_shrink_cor(R, n)
}
\arguments{
\item{R}{Square correlation matrix (p x p)}

\item{n}{Sample size used to compute R}
}
\value{
Shrunk correlation matrix of same dimensions as R
}
\description{
Applies Ledoit-Wolf shrinkage to a pre-computed sample correlation matrix
using the OAS formula (Chen et al. 2010). Shrinkage target is the
identity matrix.
}
