% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigQTL.R
\name{run_allbutone}
\alias{run_allbutone}
\title{Run all-but-one conditioning for a phenotype}
\usage{
run_allbutone(
  conditioning_snps,
  stepwise_tables,
  bigsnp,
  y,
  snp_indices,
  cis_snps_pheno,
  design_base,
  ind.row.snp,
  do_allbutone,
  ncores,
  verbose = FALSE
)
}
\arguments{
\item{conditioning_snps}{Character vector of independent SNP names}

\item{stepwise_tables}{List of data frames from stepwise conditioning}

\item{bigsnp}{bigSNP object}

\item{y}{Numeric phenotype vector}

\item{snp_indices}{Integer vector of cis-SNP column indices}

\item{cis_snps_pheno}{Character vector of cis-SNP names}

\item{design_base}{Data frame of covariates}

\item{ind.row.snp}{Integer vector of row indices for samples in genotype FBM}

\item{do_allbutone}{Logical; perform all-but-one conditioning}

\item{ncores}{Number of cores}

\item{verbose}{Logical; print progress messages (default FALSE)}
}
\value{
List of data frames with all-but-one results
}
\description{
Run all-but-one conditioning for a phenotype
}
\keyword{internal}
